Fix docs and reduce redundancy in implementation
- The documentation for the dynamic icon case was incorrect; It used the
form `(KIND ICON-FN)` as an example when code actually expected
`(KIND :fn ICON-FN :face FACE)`.
- The implementation of dynamic icons was single-cased outside of
`nerd-icons-corfu--get-by-kind`, which led to code duplication and
non-intuitive implementation. This has been fixed.
- The need for a t KIND was removed, with ? as the default, but it can
still be added to replace ? as the fallback symbol. Because of this, the
default for the variable `nerd-icons-corfu-mapping` was changed to not
include KIND t.